Moved the submitForm and invalidForm to the utils file.
